# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertOut": {
        "properties": {
          "areas": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Areas",
            "type": "array"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "fetched_at": {
            "title": "Fetched At",
            "type": "string"
          },
          "freshness": {
            "enum": [
              "FRESH",
              "RECENT",
              "STALE"
            ],
            "title": "Freshness",
            "type": "string"
          },
          "geo_match": {
            "anyOf": [
              {
                "enum": [
                  "voivodeship",
                  "unresolved"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geo Match"
          },
          "issuing_office": {
            "title": "Issuing Office",
            "type": "string"
          },
          "probability_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability Pct"
          },
          "published_at": {
            "title": "Published At",
            "type": "string"
          },
          "severity_raw": {
            "title": "Severity Raw",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "valid_from": {
            "title": "Valid From",
            "type": "string"
          },
          "valid_until": {
            "title": "Valid Until",
            "type": "string"
          }
        },
        "required": [
          "external_id",
          "source",
          "event_type",
          "severity_raw",
          "probability_pct",
          "issuing_office",
          "description",
          "comment",
          "areas",
          "valid_from",
          "valid_until",
          "published_at",
          "fetched_at",
          "freshness",
          "geo_match"
        ],
        "title": "AlertOut",
        "type": "object"
      },
...
      "DashboardAir": {
        "properties": {
          "assignment_method": {
            "title": "Assignment Method",
            "type": "string"
          },
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "coverage": {
            "enum": [
              "exact",
              "nearby",
              "regional"
            ],
            "title": "Coverage",
            "type": "string"
          },
          "coverage_radius_km": {
            "title": "Coverage Radius Km",
            "type": "integer"
          },
          "distance_km": {
            "title": "Distance Km",
            "type": "number"
          },
          "index": {
            "$ref": "#/components/schemas/AirIndex"
          },
          "observed_at": {
            "title": "Observed At",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AirParam"
            },
            "title": "Params",
            "type": "object"
          },
          "source": {
            "const": "gios",
            "title": "Source",
            "type": "string"
          },
          "source_status": {
            "$ref": "#/components/schemas/SourceStatusOut"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "station_name": {
            "title": "Station Name",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "station_name",
          "source",
          "attribution",
          "observed_at",
          "params",
          "index",
          "distance_km",
          "assignment_method",
          "coverage",
          "coverage_radius_km",
          "source_status"
        ],
        "title": "DashboardAir",
        "type": "object"
      },
      "DashboardAlerts": {
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Items",
            "type": "array"
          },
          "scope": {
            "const": "national",
            "title": "Scope",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_status": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SourceStatusOut"
            },
            "title": "Source Status",
            "type": "object"
          }
        },
        "required": [
          "scope",
          "source",
          "attribution",
          "items",
          "source_status"
        ],
        "title": "DashboardAlerts",
        "type": "object"
      },
      "DashboardArea": {
        "properties": {
          "air": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashboardAir"
              },
              {
                "type": "null"
              }
            ]
          },
          "coverage": {
            "$ref": "#/components/schemas/DashboardCoverage"
          },
          "forecast": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashboardForecast"
              },
              {
                "type": "null"
              }
            ]
          },
          "geo_area_id": {
            "title": "Geo Area Id",
            "type": "integer"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "local_alerts": {
            "items": {
              "$ref": "#/components/schemas/AlertOut"
            },
            "title": "Local Alerts",
            "type": "array"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outdoor": {
            "$ref": "#/components/schemas/DashboardOutdoor"
          },
          "pollen": {
            "$ref": "#/components/schemas/DashboardPollen"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "weather": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DashboardWeather"
              },
              {
                "type": "null"
              }
            ]
          },
          "weather_polling_active": {
            "title": "Weather Polling Active",
            "type": "boolean"
          }
        },
        "required": [
          "geo_area_id",
          "slug",
          "name",
          "latitude",
          "longitude",
          "weather_polling_active",
          "air",
          "weather",
          "forecast",
          "outdoor",
          "pollen",
          "coverage",
          "local_alerts"
        ],
        "title": "DashboardArea",
        "type": "object"
      },
...
      "DashboardResponse": {
        "properties": {
          "alerts": {
            "$ref": "#/components/schemas/DashboardAlerts"
          },
          "areas": {
            "items": {
              "$ref": "#/components/schemas/DashboardArea"
            },
            "title": "Areas",
            "type": "array"
          },
          "source_status": {
            "$ref": "#/components/schemas/DashboardSourceStatus"
          }
        },
        "required": [
          "areas",
          "alerts",
          "source_status"
        ],
        "title": "DashboardResponse",
        "type": "object"
      },
      "DashboardSourceStatus": {
        "description": "Per source, outside the nullable per-area blocks (`air`/`weather` are null without a\nstation/snapshot - \"never succeeded\" must stay distinguishable there).",
        "properties": {
          "air": {
            "$ref": "#/components/schemas/SourceStatusOut"
          },
          "weather": {
            "$ref": "#/components/schemas/SourceStatusOut"
          }
        },
        "required": [
          "air",
          "weather"
        ],
        "title": "DashboardSourceStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SourceStatusOut": {
        "properties": {
          "freshness": {
            "enum": [
              "FRESH",
              "RECENT",
              "STALE",
              "UNAVAILABLE"
            ],
            "title": "Freshness",
            "type": "string"
          },
          "last_success_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          }
        },
        "required": [
          "freshness",
          "last_success_at"
        ],
        "title": "SourceStatusOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Za Oknem API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dashboard/latest": {
      "get": {
        "description": "Combined per-location view: weather (per geo_area, always) + nearest GIOŚ\nstation's full param set (within REGIONAL_MAX_KM, classified by ADR-029 `coverage`;\nADR-006 nearest-station join, not a general geo engine). Reads only from our own DB\n(rule #14); the join happens in Python - three small independent result sets,\nnot worth a cross-table SQL join.\n\nTASK-4.1: nearest station now carries every ingested param (PM2.5/PM10/NO2/\nSO2/O3/CO/C6H6), not just PM2.5 — same `params` dict shape as /air/latest.\n\nTASK-6.2(8) / ADR-026: `?geo_area_id=N` narrows `areas` to that one area (any known\narea, also one without active polling - it then carries `weather_polling_active=false`\nand empty blocks instead of vanishing); unknown id = 404. Without it: the actively\npolled areas, as before (imported gminas must not fan out into ~2.5k areas).",
        "operationId": "dashboard_latest_api_v1_dashboard_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "geo_area_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 2147483647,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Geo Area Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dashboard Latest"
      }
    },
...
  }
}